task.updateAction for end-user actions
call task.updateAction from controller
then use Task.update only for updating model attributes
avoid complex logic in lifecycle callbacks
move some Task tests in separate files
created .x.whatever.test.js naming conventions
so that the basic model tests get executed before specific methods
