+++ test/integration/models/Task.test.js
@@ -52,38 +52,6 @@ describe('Task model', function() {
     });
 
   });
-  describe('draft task', function() {
-    beforeEach(function(done) {
-      Task.create(taskFixtures.draft).then(function(newTask) {
-        task = newTask;
-        done();
-      })
-      .catch(done);
-    });
-    it('submitted sets submittedAt', function(done) {
-      Task.update(task.id, {id: task.id, state: 'submitted'}).exec(function (err, updated_task) {
-        if ( err ) { return done( err ); }
-        assert(updated_task[0].submittedAt <= new Date(), 'submittedAt not set');
-        done();
-      });
-    });
-  });
-  describe('submitted task', function() {
-    beforeEach(function(done) {
-      Task.create(taskFixtures.submitted).then(function(newTask) {
-        task = newTask;
-        done();
-      })
-      .catch(done);
-    });
-    it('open sets publishedAt', function(done) {
-      Task.update(task.id, {id: task.id, state: 'open'}).exec(function (err, updated_task) {
-        if ( err ) { return done( err ); }
-        assert(updated_task[0].publishedAt <= new Date(), 'publishedAt not set');
-        done();
-      });
-    });
-  });
 
   describe('.authorized', function() {
     describe('without user', function() {
@@ -198,103 +166,4 @@ describe('Task model', function() {
     });
 
   });
-  describe('#byCategory', function() {
-    describe('with no tasks', function() {
-      it('returns correct categories', function(done) {
-        Task.byCategory().then(function(result) {
-          assert.deepEqual(result, {
-            assigned: [],
-            completed: [],
-            draft: [],
-            open: [],
-            submitted: []
-          });
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-      });
-    });
-    describe('with draft task', function() {
-      beforeEach(function(done) {
-        User.create(userFixtures.minAttrs).then(function(user) {
-          return Task.create(taskFixtures.draft);
-        }).then(function(tasks) {
-            done();
-        }).catch(function(err) {
-          done(err);
-        })
-      });
-      it('returns task correctly', function(done) {
-        Task.byCategory().then(function(result) {
-          assert.equal(result.draft.length, 1);
-          assert.equal(result.draft[0].state, 'draft');
-          assert.isNotNull(result.draft[0].user);
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-      });
-
-    });
-    describe('with signups', function() {
-      var users, tasks;
-      beforeEach(function(done) {
-        users = [userFixtures.minAttrs, userFixtures.emmy, userFixtures.alan]
-        User.create(users).then(function(users) {
-          return Task.create([taskFixtures.open, taskFixtures.assigned  ]);
-        }).then(function(tasks) {
-            done();
-        }).catch(function(err) {
-          done(err);
-        })
-      });
-      it('returns task correctly', function(done) {
-        Task.byCategory().then(function(result) {
-          assert.equal(result.open.length, 1);
-          assert.equal(result.assigned.length, 1);
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-      });
-    });
-    describe('with tasks in each category', function() {
-      beforeEach(function(done) {
-        User.create(userFixtures.minAttrs).then(function(user) {
-          var keys = ['draft', 'submitted', 'open',
-                               'assigned', 'completed', 'archived'];
-          attrList = _.pick(taskFixtures, keys);
-          attrList = _.values(attrList);
-          return Task.create(attrList);
-        }).then(function(tasks) {
-            done();
-        }).catch(function(err) {
-          done(err);
-        })
-      });
-      it('returns tasks in correct categories', function(done) {
-        Task.byCategory().then(function(result) {
-          assert.equal(result.draft.length, 1);
-          assert.equal(result.draft[0].state, 'draft');
-          assert.equal(result.submitted.length, 1);
-          assert.equal(result.submitted[0].state, 'submitted');
-          assert.equal(result.open.length, 1);
-          assert.equal(result.open[0].state, 'open');
-          assert.equal(result.assigned.length, 1);
-          assert.equal(result.assigned[0].state, 'assigned');
-          assert.equal(result.completed.length, 1);
-          assert.equal(result.completed[0].state, 'completed');
-          done();
-        })
-        .catch(function (err) {
-          done(err);
-        });
-      });
-    });
-  });
-
 });
